+++ CHANGELOG.md
@@ -233,3 +233,4 @@
 - 2023-12-31: Test deal, daemon, README, bettercontact, +3
 - 2024-01-01: Docs sender, conf, users, find email, +3
 - 2024-01-02: Bench contacts, find email, leads, lead, +4
+- 2024-01-03: Init task, lib, follow up, campaigns, +3